Rename file for basic kernels
