+++ baybe/kernels/__init__.py
@@ -1,5 +1,5 @@
 """Kernels for Gaussian process surrogate models."""
 
-from baybe.kernels.kernels import MaternKernel
+from baybe.kernels.basic import MaternKernel
 
 __all__ = ["MaternKernel"]
